Reorganized structure of the project
