--- src/com/bad_java/homework/Homework.java
@@ -1,1 +1,0 @@
-Homework
