+++ src/com/bad_java/lectures/_02/GitExample.java
@@ -1,0 +1,8 @@
+package com.bad_java.lectures._02;
+
+public class GitExample {
+
+  public static void main(String[] args) {
+    System.out.println("Git example!");
+  }
+}
